Fixed a problem parsing redis message when keyPrefix includes :
In a webfarm, the pubsub mechanism ensures cache is cleared on each node in the
farm. If keyPrefix configuration options includes a ':', a bug in the parsing
means the cache will never be cleared on the other machines.

The fix involves a change to how the instanceUniqueId, type and data are parsed
out of the message.  The last two entries are always considered to be the type
and data, respectively and the instanceUniqueId check changes to be the start
of the message.

Fixes #17